chore: move integration tests from unit
